use 64bit immediate integers

+++ factory/cf_factory.h
@@ -20,14 +20,14 @@ private:
 public:
     static int gettype () { return currenttype; }
     static void settype ( int type );
-    static InternalCF * basic ( int value );
-    static InternalCF * basic ( int type, int value );
+    static InternalCF * basic ( long value );
+    static InternalCF * basic ( int type, long value );
     static InternalCF * basic ( const char * str );
     static InternalCF * basic ( const char * str, int base );
     static InternalCF * basic ( int type, const char * const str );
-    static InternalCF * basic ( int type, int value, bool nonimm );
+    static InternalCF * basic ( int type, long value, bool nonimm );
     static InternalCF * basic ( const mpz_ptr num );
-    static InternalCF * rational ( int num, int den );
+    static InternalCF * rational ( long num, long den );
     static InternalCF * rational ( const mpz_ptr num, const mpz_ptr den, bool normalize );
     static InternalCF * poly ( const Variable & v, int exp, const CanonicalForm & c );
     static InternalCF * poly ( const Variable & v, int exp = 1 );
